Move pdf files to new package
