+++ src/integrationTest/java/uk/gov/hmcts/reform/sendletter/services/PdfCreatorTest.java
@@ -3,9 +3,8 @@ package uk.gov.hmcts.reform.sendletter.services;
 import org.junit.Before;
 import org.junit.Test;
 import uk.gov.hmcts.reform.pdf.generator.HTMLToPDFConverter;
-import uk.gov.hmcts.reform.slc.services.steps.getpdf.PdfCreator;
-import uk.gov.hmcts.reform.slc.services.steps.getpdf.duplex.DuplexPreparator;
-import uk.gov.hmcts.reform.slc.services.steps.getpdf.exceptions.InvalidPdfException;
+import uk.gov.hmcts.reform.sendletter.exception.InvalidPdfException;
+import uk.gov.hmcts.reform.sendletter.services.util.DuplexPreparator;
 
 import java.util.Base64;
 import java.util.List;
