+++ src/main/java/uk/gov/hmcts/reform/sendletter/exception/PdfMergeException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.slc.services.steps.getpdf.exceptions;
+package uk.gov.hmcts.reform.sendletter.exception;
 
 public class PdfMergeException extends RuntimeException {
     public PdfMergeException(String message, Throwable cause) {
