+++ src/main/java/uk/gov/hmcts/reform/sendletter/services/PdfCreator.java
@@ -1,10 +1,11 @@
-package uk.gov.hmcts.reform.slc.services.steps.getpdf;
+package uk.gov.hmcts.reform.sendletter.services;
 
 import org.apache.http.util.Asserts;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.reform.sendletter.exception.InvalidPdfException;
 import uk.gov.hmcts.reform.sendletter.model.in.Document;
-import uk.gov.hmcts.reform.slc.services.steps.getpdf.duplex.DuplexPreparator;
-import uk.gov.hmcts.reform.slc.services.steps.getpdf.exceptions.InvalidPdfException;
+import uk.gov.hmcts.reform.sendletter.services.util.DuplexPreparator;
+import uk.gov.hmcts.reform.sendletter.services.util.IHtmlToPdfConverter;
 
 import java.util.Base64;
 import java.util.List;
@@ -54,6 +55,7 @@ public class PdfCreator {
         }
     }
 
+    // TODO dependency?
     private byte[] decodePdf(String base64encodedPdf) {
         try {
             return Base64.getDecoder().decode(base64encodedPdf);
