+++ src/test/java/uk/gov/hmcts/reform/sendletter/services/DuplexPreparatorTest.java
@@ -1,7 +1,9 @@
-package uk.gov.hmcts.reform.slc.services.steps.getpdf.duplex;
+package uk.gov.hmcts.reform.sendletter.services;
 
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.junit.Test;
+import uk.gov.hmcts.reform.sendletter.exception.DuplexException;
+import uk.gov.hmcts.reform.sendletter.services.util.DuplexPreparator;
 
 import static com.google.common.io.Resources.getResource;
 import static com.google.common.io.Resources.toByteArray;
